% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboutChristina.R
\name{aboutChristina}
\alias{aboutChristina}
\title{The aboutChristina Function}
\usage{
aboutChristina()
}
\value{
NA
}
\description{
This function contains 1 lie and 4 truths about my loved ones and I. The lie is #1. I am a really terrible dancer.
}
